- fixes missing documentation and deprecation information on functions
